Pull VizMovingHead out of Visualizer.
